[REF](Examples): update pipeline example names and descriptions; rename permit pipeline example typo and resolve Create and Permit displaying example
