+++ workspaces/azure-devops/plugins/scaffolder-backend-module-azure-devops/src/actions/devopsPermitPipeline.examples.ts
